# docs/api/openapi.m13.json
{
...
  "components": {
    "schemas": {
...
      "CapabilityFailureDto": {
        "additionalProperties": false,
        "properties": {
          "capability": {
            "title": "Capability",
            "type": "string"
          },
          "error_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Category"
          },
          "error_message_redacted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message Redacted"
          }
        },
        "required": [
          "capability"
        ],
        "title": "CapabilityFailureDto",
        "type": "object"
      },
...
      "CompatibilityViewDto": {
        "additionalProperties": false,
        "description": "模型兼容性视图：声明能力 + endpoint 摘要（不含未探测的推断）。",
        "properties": {
          "endpoint_healthy_hint": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint Healthy Hint"
          },
          "endpoint_id": {
            "title": "Endpoint Id",
            "type": "string"
          },
          "hard_capability_requirements": {
            "items": {
              "type": "string"
            },
            "title": "Hard Capability Requirements",
            "type": "array"
          },
          "model": {
            "$ref": "#/components/schemas/ModelReadDto"
          }
        },
        "required": [
          "model",
          "endpoint_id"
        ],
        "title": "CompatibilityViewDto",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelCreateDto": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "additionalProperties": false,
            "title": "Capabilities",
            "type": "object"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "endpoint_id": {
            "minLength": 1,
            "title": "Endpoint Id",
            "type": "string"
          },
          "model_name": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          }
        },
        "required": [
          "endpoint_id",
          "model_name"
        ],
        "title": "ModelCreateDto",
        "type": "object"
      },
      "ModelReadDto": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "additionalProperties": false,
            "title": "Capabilities",
            "type": "object"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "endpoint_id": {
            "title": "Endpoint Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "version": {
            "description": "resource version（ETag 值，If-Match 用）",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "endpoint_id",
          "model_name",
          "display_name",
          "enabled",
          "capabilities",
          "version"
        ],
        "title": "ModelReadDto",
        "type": "object"
      },
      "ModelRuntimeFingerprintDto": {
        "additionalProperties": false,
        "description": "AGENTS.md §4 漂移可见性字段（全部脱敏，无 secret）。",
        "properties": {
          "endpoint_config_digest": {
            "title": "Endpoint Config Digest",
            "type": "string"
          },
          "observed_capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Observed Capabilities",
            "type": "array"
          },
          "probe_suite_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probe Suite Digest"
          },
          "returned_model_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Model Identifier"
          },
          "system_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Fingerprint"
          }
        },
        "required": [
          "endpoint_config_digest"
        ],
        "title": "ModelRuntimeFingerprintDto",
        "type": "object"
      },
      "ModelUpdateDto": {
        "additionalProperties": false,
        "description": "更新模型：全部可选；capabilities 整体替换（含 PROBED 断言）。",
        "properties": {
          "capabilities": {
            "anyOf": [
              {
                "additionalProperties": false,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capabilities"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "title": "ModelUpdateDto",
        "type": "object"
      },
...
      "ProbeResultDto": {
        "additionalProperties": false,
        "description": "capability probe 结果：system_fingerprint=None 时前端必须渲染\n'configuration reproducible / provider fingerprint unavailable'，\n禁止美化为一律 fully reproducible。",
        "properties": {
          "capability_failures": {
            "items": {
              "$ref": "#/components/schemas/CapabilityFailureDto"
            },
            "title": "Capability Failures",
            "type": "array"
          },
          "error_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Category"
          },
          "error_message_redacted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message Redacted"
          },
          "fingerprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelRuntimeFingerprintDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "observed_capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Observed Capabilities",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "probed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probed At"
          },
          "provider_fingerprint_available": {
            "title": "Provider Fingerprint Available",
            "type": "boolean"
          },
          "returned_model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Model Name"
          },
          "system_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Fingerprint"
          }
        },
        "required": [
          "model_id",
          "ok",
          "provider_fingerprint_available"
        ],
        "title": "ProbeResultDto",
        "type": "object"
      },
...
      "ValidationError": {
        "additionalProperties": false,
        "properties": {
          "ctx": {
            "additionalProperties": false,
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Research Console 控制面：配置 / 探测 / 运行 / 审批 / 审计",
    "title": "Research OS Control Plane API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/models": {
      "get": {
        "operationId": "list_models_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "endpoint_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelReadDto"
                  },
                  "title": "Response List Models Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      },
      "post": {
        "operationId": "create_model_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelReadDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}": {
      "get": {
        "operationId": "get_model_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelReadDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "models"
        ]
      },
      "patch": {
        "operationId": "update_model_models__model_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelReadDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}/compatibility": {
      "get": {
        "description": "兼容性视图：声明能力（含来源标注）+ endpoint 摘要；不含未探测推断。",
        "operationId": "model_compatibility_models__model_id__compatibility_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatibilityViewDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Compatibility",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}/probe": {
      "post": {
        "description": "执行正式 capability probe（run_probe + fingerprint，M12 语义）。\n\n成功后把 PROBED 断言合并进 ModelDefinition 并持久化（probe 是配置\n操作；运行后能力声明与 backend truth 一致）。",
        "operationId": "probe_model_models__model_id__probe_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProbeResultDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Probe Model",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}